Complex Types (#15)
* Add collection handling, uuid type, and basic logging

* Deserialize referenced types, fix tests

* Add validator, implement alias support, improve swagger writter performance

* Enhance validator, write complex types swagger

* Remove reduntant alias validation

* Remove unused definitions

Complex Types have little difference with entityTypes metadata-wise, so it makes much more sense to use the same code behavior as entityTypes for the output of their definitions

* Implement Enums, fix validator, fix binary type

* Fix format

* Implement BFS complex type exploration

* Fix bug causing complex type collections to not be output properly

* Remove unused parameter required

Required properties are already handled by the sidecar configuration.

* Remove unused security metadata

* Revert singleton design pattern

The abuse of singletons in the project led to an antipattern making it hard to test and unmaintainable in the long term. Writting long function signatures and calls is preferable than having to implement magic mocks and unsupported tricks in jest for supporting Singletons in Typescript. It's also important to remark that singletons are not natively supported by typescript and Instances would be handled within the classes, causing problems with scoping and calling from different threads or processes.

* Add testing for definitions and validator, expand coverage for existing cases

* Remove sentinels from enums

* Implement testing for collections and enums

* Implement feedback

* Remove unnecessary initialization, change error msg